feat: migrate gix-cmp Collapsible, KeyValuePair and KeyValuePairInfo (#7169)
# Motivation

We **need** to integrate the latest version of **Gix-cmp**, which
includes the migration of `Collapsible`, `KeyInfoPair`, and
`KeyInfoPairInfo` to **Svelte v5**. Otherwise, the collapsible element
may eventually stop working as expected, as it has become incompatible
with the latest version of Svelte.

# Changes in Gix-cmp

- https://github.com/dfinity/gix-components/pull/696
- https://github.com/dfinity/gix-components/pull/698

# Notes

Unfortunately not really a way to size down the PR 😢. Collapsible,
KeyValuePair, KeyValuePairInfo and Section are related and using each
others so because of the slot migration to snippet, we cannot just
migrate one or the other.

# Changes

- Bump `gix-cmp@next`
- Migrate required slots to snippets.
- Added few guards as snippet did not inherit their top nullish guards
- Migrate `ProposalSystemInfoEntry` to Svelte v5 to avoid naming clashes
on `value`.
- Migrate `ExpandableProposalNeurons` and `StakeNeuronToVote` to Svelte
v5 to use new `Collapsible`.

# Tests

- Added an `await tick()` in the failing tests to ensure UI updates are
applied (similar to the approach used here:
https://github.com/dfinity/nns-dapp/pull/6922).

---------

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>
Co-authored-by: Max Strasinsky <maksims.strasinskis@dfinity.org>
Co-authored-by: Max Strasinsky <98811342+mstrasinskis@users.noreply.github.com>