fix(aggregator): handle new InfoTopic extension field (#7324)
# Motivation

The latest upgrade of candid files in
[ic-js](https://github.com/dfinity/ic-js/pull/1032/files#diff-684ab36c5e16fec984732ec9c53e9fe5690f4dddf1ddd713d7fd48a210de5707R1051)
introduced a new field, `extension_operations`, to the `InfoTopic`
field. This change requires that the nns-dapp be updated to initialize
this new field. Since the front end does not need it and the value is
not exposed by the Aggregator, this PR initializes it to an empty value,
`[]`.

NOTE: If the value is needed in the future, a proper adaptation will be
made at that time.

# Changes

- Add `extension_operations` field
- Initialize new property to `[]`

# Tests

- CI should pass as before.

# Todos

- [x] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed?

---------

Co-authored-by: gix-bot <gix-bot@users.noreply.github.com>