feat(reports): add transaction metadata to each row (#7369)
# Motivation

The generated report includes a metadata section with information about
the `neuronId` and `transactionId`. However, this information may be
lost if the user filters by any field. This PR duplicates this
information into each transaction row, ensuring that users can always
see the field.

Additionally, this PR makes some related changes, such as renaming the
files to include the period used for their generation and changing the
`TX Id` column to `Index` to align with the notation used in the
Dashboard.

[NNS1-4133](https://dfinity.atlassian.net/browse/NNS1-4133)

# Changes

* Add `neuronId` and `accountId` to each transaction row.
* Add suffix to file name to include the transaction period.
* Renaming some labels for clarity.

# Tests

- Update tests to check for new columns and new labels.

# Todos

- [x] Accessibility (a11y) – any impact?
- [x] Changelog – is it needed?


[NNS1-4133]:
https://dfinity.atlassian.net/browse/NNS1-4133?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ